+++ autotest/config.yaml
@@ -21,6 +21,7 @@ tp_config:
 
 turbomind_chat_model:
     - meta-llama/Llama-3.2-1B-Instruct
+    - meta-llama/Llama-3.2-3B-Instruct
     - meta-llama/Meta-Llama-3-1-8B-Instruct
     - meta-llama/Meta-Llama-3-1-8B-Instruct-AWQ
     - meta-llama/Meta-Llama-3-8B-Instruct
@@ -51,6 +52,7 @@ turbomind_chat_model:
     - Qwen/Qwen2-7B-Instruct-GPTQ-Int4
     - Qwen/Qwen2-57B-A14B-Instruct-GPTQ-Int4
     - mistralai/Mistral-7B-Instruct-v0.3
+    - mistralai/Mistral-Nemo-Instruct-2407
     - mistralai/Mixtral-8x7B-Instruct-v0.1
     - lmdeploy/llama2-chat-7b-w4
     - baichuan-inc/Baichuan2-7B-Chat
@@ -69,11 +71,14 @@ pytorch_chat_model:
     - meta-llama/Meta-Llama-3-8B-Instruct
     - meta-llama/Meta-Llama-3-1-8B-Instruct
     - meta-llama/Llama-3.2-1B-Instruct
+    - meta-llama/Llama-3.2-3B-Instruct
+    - meta-llama/Llama-3.2-11B-Vision-Instruct
     - meta-llama/Llama-2-7b-chat-hf
     - internlm/internlm2_5-7b-chat
     - internlm/internlm2_5-20b-chat
     - internlm/internlm2-chat-20b
     - internlm/internlm-chat-20b
+    - OpenGVLab/InternVL2-1B
     - OpenGVLab/InternVL2-2B
     - OpenGVLab/InternVL2-4B
     - OpenGVLab/InternVL2-8B
@@ -106,6 +111,43 @@ pytorch_chat_model:
     - microsoft/Phi-3-mini-4k-instruct
     - microsoft/Phi-3-vision-128k-instruct
 
+turbomind_vl_model:
+    - Qwen/Qwen-VL-Chat
+    - liuhaotian/llava-v1.5-13b
+    - liuhaotian/llava-v1.6-vicuna-7b
+    - 01-ai/Yi-VL-6B
+    - deepseek-ai/deepseek-vl-1.3b-chat
+    - OpenGVLab/InternVL-Chat-V1-5
+    - OpenGVLab/Mini-InternVL-Chat-2B-V1-5
+    - OpenGVLab/InternVL2-1B
+    - OpenGVLab/InternVL2-2B
+    - OpenGVLab/InternVL2-8B
+    - OpenGVLab/InternVL2-26B
+    - OpenGVLab/InternVL2-40B
+    - internlm/internlm-xcomposer2d5-7b
+    - internlm/internlm-xcomposer2-4khd-7b
+    - openbmb/MiniCPM-Llama3-V-2_5
+    - openbmb/MiniCPM-V-2_6
+
+pytorch_vl_model:
+    - meta-llama/Llama-3.2-11B-Vision-Instruct
+    - OpenGVLab/InternVL-Chat-V1-5
+    - OpenGVLab/Mini-InternVL-Chat-2B-V1-5
+    - OpenGVLab/InternVL2-1B
+    - OpenGVLab/InternVL2-2B
+    - OpenGVLab/InternVL2-4B
+    - OpenGVLab/InternVL2-8B
+    - OpenGVLab/InternVL2-26B
+    - OpenGVLab/InternVL2-40B
+    - OpenGVLab/Mono-InternVL-2B
+    - Qwen/Qwen2-VL-2B-Instruct
+    - Qwen/Qwen2-VL-7B-Instruct
+    - THUDM/cogvlm-chat-hf
+    - THUDM/cogvlm2-llama3-chinese-chat-19B
+    - THUDM/glm-4v-9b
+    - microsoft/Phi-3-vision-128k-instruct
+    - microsoft/Phi-3.5-vision-instruct
+
 turbomind_base_model:
     - internlm/internlm2_5-7b
     - internlm/internlm2_5-1_8b
@@ -118,33 +160,6 @@ pytorch_base_model:
     - internlm/internlm2_5-1_8b
     - internlm/internlm2_5-20b
     - bigcode/starcoder2-7b
-
-vl_model:
-    - Qwen/Qwen-VL-Chat
-    - liuhaotian/llava-v1.5-13b
-    - liuhaotian/llava-v1.6-vicuna-7b
-    - 01-ai/Yi-VL-6B
-    - deepseek-ai/deepseek-vl-1.3b-chat
-    - OpenGVLab/InternVL-Chat-V1-5
-    - OpenGVLab/Mini-InternVL-Chat-2B-V1-5
-    - OpenGVLab/InternVL2-1B
-    - OpenGVLab/InternVL2-2B
-    - OpenGVLab/InternVL2-4B
-    - OpenGVLab/InternVL2-8B
-    - OpenGVLab/InternVL2-26B
-    - OpenGVLab/InternVL2-40B
-    - OpenGVLab/Mono-InternVL-2B
-    - Qwen/Qwen2-VL-2B-Instruct
-    - Qwen/Qwen2-VL-7B-Instruct
-    - internlm/internlm-xcomposer2d5-7b
-    - internlm/internlm-xcomposer2-4khd-7b
-    - THUDM/cogvlm-chat-hf
-    - THUDM/cogvlm2-llama3-chinese-chat-19B
-    - THUDM/glm-4v-9b
-    - microsoft/Phi-3.5-vision-instruct
-    - microsoft/Phi-3-vision-128k-instruct
-    - openbmb/MiniCPM-Llama3-V-2_5
-    - openbmb/MiniCPM-V-2_6
 
 turbomind_quatization:
     no_awq:
@@ -184,10 +199,12 @@ pytorch_quatization:
         - internlm/internlm2_5-20b
         - internlm/internlm2_5-7b
     no_kvint4:
+        - OpenGVLab/InternVL2-1B
         - OpenGVLab/InternVL2-4B
         - deepseek-ai/DeepSeek-V2-Lite-Chat
         - microsoft/Phi-3-mini-4k-instruct
         - microsoft/Phi-3-vision-128k-instruct
+        - microsoft/Phi-3.5-vision-instruct
         - openbmb/MiniCPM-V-2_6
     no_kvint8:
         - deepseek-ai/DeepSeek-V2-Lite-Chat
