+++ autotest/tools/pipeline/test_pipeline_chat_turbomind_llm.py
